fix: checkStuck wouldn't work if duration between RequestButtons were longer than stuckDurationlimit

+++ timer/timer.go
@@ -7,7 +7,7 @@ import (
 var timer *time.Timer
 var timerActive bool
 
-func TimerStart(duration float64) {
+func TimerStart(duration int64) {
 	timer = time.NewTimer(time.Duration(duration) * time.Second)
 	timerActive = true
 }
